Create Dashboard-1.xlsm
The first of the them is the sales dashboard, which contains a pivot table with fields for sales persons, months and years, the overall sales amounts and the possibility to add interactive slicers based on executive, region or product. It supports the analysis of sales activity, reveals tendencies, and facilitates managers’ decision making to promote improved sales performance.